feat: Implement chat session handler with JWT token generation and secure cookie management
