+++ server/server.js
@@ -1,4 +1,4 @@
-import dbDeleteExpertEvalHandler from '../api/db/db-delete-expert-eval.js';
+﻿import dbDeleteExpertEvalHandler from '../api/db/db-delete-expert-eval.js';
 import checkUrlHandler from '../api/util/util-check-url.js';
 import similarChatsHandler from '../api/vector/vector-similar-chats.js';
 import express from 'express';
@@ -21,7 +21,7 @@ import dbBatchDeleteHandler from '../api/batch/batch-delete.js';
 import batchesDeleteAllHandler from '../api/batch/batches-delete-all.js';
 import anthropicContextAgentHandler from '../api/anthropic/anthropic-context.js';
 import openAIContextAgentHandler from '../api/openai/openai-context.js';
-import dbChatSessionHandler from '../api/db/db-chat-session.js';
+import chatSessionHandler from '../api/chat/chat-session.js';
 import chatSimilarAnswerHandler from '../api/chat/chat-similar-answer.js';
 import chatPIICheckHandler from '../api/chat/chat-pii-check.js';
 import chatDetectLanguageHandler from '../api/chat/chat-detect-language.js';
@@ -118,7 +118,7 @@ app.post('/api/feedback/feedback-get-expert', feedbackGetExpertHandler);
 app.post('/api/feedback/feedback-get-public', feedbackGetPublicHandler);
 app.post('/api/feedback/feedback-delete-expert', feedbackDeleteExpertHandler);
 app.post('/api/db/db-persist-interaction', dbPersistInteraction);
-app.get('/api/db/db-chat-session', dbChatSessionHandler);
+app.get('/api/chat/chat-session', chatSessionHandler);
 app.get('/api/db/db-verify-chat-session', dbVerifyChatSessionHandler);
 app.get('/api/batch/batch-list', dbBatchListHandler);
 app.get('/api/batch/batch-retrieve', dbBatchRetrieveHandler);
